add models/testScore; improve testInstance syncing

+++ backend/models/testQuestion.js
@@ -1,9 +1,9 @@
 module.exports = function(sequelize, DataTypes) {
     var TestQuestion = sequelize.define('TestQuestion', {
         number: DataTypes.INTEGER,
-        maxPoints: {type: DataTypes.FLOAT, field: 'max_points'},
-        pointsList: {type: DataTypes.ARRAY(DataTypes.FLOAT), field: 'points_list'},
-        initPoints: {type: DataTypes.FLOAT, field: 'init_points'},
+        maxPoints: {type: DataTypes.DOUBLE, field: 'max_points'},
+        pointsList: {type: DataTypes.ARRAY(DataTypes.DOUBLE), field: 'points_list'},
+        initPoints: {type: DataTypes.DOUBLE, field: 'init_points'},
         testId: {type: DataTypes.INTEGER, field: 'test_id'},
         zoneId: {type: DataTypes.INTEGER, field: 'zone_id'},
         questionId: {type: DataTypes.INTEGER, field: 'question_id'},
